fix(core): Upgrade core to new version with mwei gasprice, rounding up wants, and 127 bits for gives (#72)
* fix(core): Upgrade core to new version with mwei gasprice, rounding up wants, and 127 bits for gives

* bump core a bit more

* allow kandel gasprice to be 26 bits, and fix some checks

* use MAX_SAFE_VOLUME

* add gasprice to audit_trail

* review fixes

* upg core again

+++ test/strategies/kandel/abstract/KandelTest.t.sol
@@ -180,7 +180,7 @@ abstract contract KandelTest is StratTest {
   }
 
   function getParams(GeometricKandel aKandel) internal view returns (GeometricKandel.Params memory params) {
-    (uint16 gasprice, uint24 gasreq, uint104 stepSize, uint112 pricePoints) = aKandel.params();
+    (uint32 gasprice, uint24 gasreq, uint88 stepSize, uint112 pricePoints) = aKandel.params();
 
     params.gasprice = gasprice;
     params.gasreq = gasreq;
@@ -394,7 +394,7 @@ abstract contract KandelTest is StratTest {
     }
     GeometricKandel.Params memory params;
     params.pricePoints = uint112(pricePoints);
-    params.stepSize = uint104(stepSize);
+    params.stepSize = uint88(stepSize);
 
     kandel.populate{value: 0.1 ether}(distribution, params, 0, 0);
   }
